% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.points.R
\docType{data}
\name{animal.points}
\alias{animal.points}
\title{Example animal location data}
\format{
A data frame with 11 rows and 5 columns:
\describe{
  \item{lon.raw}{Longitude in decimal degrees}
  \item{lat.raw}{Latitude in decimal degrees}
  \item{lon.shifted}{Longitude shifted to line up with stream points to illustrate how the functions work}
  \item{lat.shifted}{Latitude shifted to line up with stream points to illustrate how the functions work}
  \item{id}{Turtle IDs}
  \item{date.time}{Dates and times of points in POSIX format}
  \item{point}{Names of each coordinate (optional)}
  }
}
\source{
Fictional example generated to demonstrate this package
}
\usage{
animal.points
}
\description{
Hypothetical data set with the GPS coordinates for two turtles
}
\keyword{datasets}
