% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.dist.R
\name{node.dist}
\alias{node.dist}
\title{Calculate minimum distance between all pairs of nodes}
\usage{
node.dist(x)
}
\arguments{
\item{x}{(data.frame) The \code{segement} object created using \code{\link{dist.per.segment}}}
}
\value{
A data frame containing the name of the first node, name of the second node, and distance between them (in meters). The diagonal of the distance matrix is not returned, and there are no duplicate pairs.
}
\description{
For a given network of nodes and distances, it will use Dijkstra’s algorithm via \code{\link[igraph:distances]{igraph}} to calculate the minimum distance between each pair of nodes following the river network.
}
\keyword{internal}
