% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Example node data}
\format{
A data frame with 8 rows and 3 columns:
\describe{
  \item{lon}{Longitude in decimal degrees}
  \item{lat}{Latitude in decimal degrees}
  \item{id}{Node names (must be integers)}
  ...
}
}
\source{
Fictional example generated to demonstrate this package
}
\usage{
nodes
}
\description{
Hypothetical node data set
}
\keyword{datasets}
