\name{buildBasisFunction}
\alias{buildBasisFunction}
\alias{buildBasisSplineMatrix}
\alias{calcArrayEffects}
\alias{calculate.nulls}
\alias{checkArguments}
\alias{edge.fit}
\alias{edge.glr}
\alias{edge.qvalue}
\alias{err.msg}
\alias{estimateDifferentialExpression}
\alias{fit.fast.model}
\alias{fit.model}
\alias{getSpanningSet}
\alias{make.ref.model.matrices}
\alias{make.snm.obj}
\alias{makeDataObject}
\alias{remove.adjustment.vars}
\alias{rm.zero.cols}
\alias{sim.function.var}
\alias{sim.intensity.dep}
\alias{sim.probe.specific}
\alias{snm.diagnostic.plot}
\title{
Internal snm functions.
}
\description{
Internal functions for the snm package not called by the user.
}
\author{
  All internal functions written by Brig Mecham <brig.mecham@sagebase.org>, except \code{edge.fit}, \code{edge.glr}, and \code{edge.qvalue} written by John D. Storey <jstorey@princeton.edu> 
}
\keyword{internal}
