% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.boundaries.R
\docType{data}
\name{stream.boundaries}
\alias{stream.boundaries}
\title{Example stream outline}
\format{
An sf data frame with 1 observation of 2 variables:
\describe{
  \item{id}{NA}
  \item{geometry}{shapefile geometry}}
}
\source{
Fictional example generated to demonstrate this package
}
\usage{
stream.boundaries
}
\description{
Hypothetical outline of example stream (polygon shapefile). Not actually needed for functions.
}
\keyword{datasets}
