% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.line.R
\docType{data}
\name{stream.line}
\alias{stream.line}
\title{Example stream network shapefile}
\format{
An "sf" data frame with 7 observations (stream segments) and 2 variables:
\describe{
  \item{id}{id of each stream segement}
  \item{geometry}{sf geometry data for each stream segement}
  }
}
\source{
Fictional example generated to demonstrate this package. Shapefile was created in QGIS and imported via st_read()
}
\usage{
stream.line
}
\description{
Shapefile (polyline) for a hypothetical stream network
}
\keyword{datasets}
